+++ src/ResourceManager/Resources/Commands.ResourceManager/Cmdlets/Implementation/Policy/PolicyHelpStrings.cs
@@ -24,7 +24,6 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation.Policy
         public const string GetPolicyAssignmentIdHelp = "The fully qualified policy assignment ID to get, including the scope, e.g. /subscriptions/{subscriptionId}/resourcegroups/{resourceGroupName}/providers/Microsoft.Authorization/policyAssignments/{policyAssignmentName}.";
         public const string GetPolicyDefinitionFilterHelp = "Limits the list of returned policy assignments to those assigning the policy definition identified by this fully qualified Id.";
         public const string GetPolicyAssignmentIncludeDescendentsHelp = "Causes the list of returned policy assignments to include all assignments related to the given scope, including those from ancestor scopes and those from descendent scopes.";
-        public const string GetPolicyAssignmentDoesNothingHelp = "This parameter is ignored if provided with -Name or -Id parameters.";
         public const string NewPolicyAssignmentNameHelp = "The name of the new policy assignment.";
         public const string NewPolicyAssignmentScopeHelp = "The scope of the new policy assignment, e.g. /providers/managementGroups/{managementGroupName}.";
         public const string NewPolicyAssignmentNotScopesHelp = "The not scopes for the new policy assignment.";
@@ -39,6 +38,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation.Policy
         public const string RemovePolicyAssignmentNameHelp = "The name of the policy assignment to delete.";
         public const string RemovePolicyAssignmentScopeHelp = "The scope of the policy assignment to delete, e.g. /providers/managementGroups/{managementGroupName}.";
         public const string RemovePolicyAssignmentIdHelp = "The fully qualified policy assignment ID to delete, including the scope, e.g. /subscriptions/{subscriptionId}/resourcegroups/{resourceGroupName}/providers/Microsoft.Authorization/policyAssignments/{policyAssignmentName}.";
+        public const string RemovePolicyAssignmentInputObjectHelp = "The policy assignment object to remove that was output from another cmdlet.";
         public const string SetPolicyAssignmentNameHelp = "The name of the policy assignment to update.";
         public const string SetPolicyAssignmentScopeHelp = "The scope of the policy assignment to update, e.g. /providers/managementGroups/{managementGroupName}.";
         public const string SetPolicyAssignmentNotScopesHelp = "The not scopes of the updated policy assignment.";
@@ -47,6 +47,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation.Policy
         public const string SetPolicyAssignmentDescriptionHelp = "The description of the updated policy assignment";
         public const string SetPolicyAssignmentMetadataHelp = "The updated metadata for the policy assignment. This can either be a path to a file name containing the metadata, or the metadata as a string.";
         public const string SetPolicyAssignmentSkuHelp = "A hash table which specifies sku properties. This parameter is deprecated and ignored.";
+        public const string SetPolicyAssignmentInputObjectHelp = "The policy assignment object to update that was output from another cmdlet.";
 
         /// <summary>
         /// Policy definition cmdlet parameter help strings
@@ -68,6 +69,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation.Policy
         public const string NewPolicyDefinitionSubscriptionIdHelp = "The subscription ID of the new policy definition.";
         public const string RemovePolicyDefinitionNameHelp = "The name of the policy definition to delete.";
         public const string RemovePolicyDefinitionIdHelp = "The fully qualified policy definition ID to delete, including the subscription or management group. e.g. /subscriptions/{subscriptionId}/providers/Microsoft.Authorization/policyDefinitions/{policyDefinitionName}.";
+        public const string RemovePolicyDefinitionInputObjectHelp = "The policy definition object to remove that was output from another cmdlet.";
         public const string ForceFlagHelp = "Do not ask for confirmation.";
         public const string RemovePolicyDefinitionManagementGroupHelp = "The name of the management group of the policy definition to delete.";
         public const string RemovePolicyDefinitionSubscriptionIdHelp = "The subscription ID of the policy definition to delete.";
@@ -80,6 +82,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation.Policy
         public const string SetPolicyDefinitionParameterHelp = "The parameters declaration of the updated policy definition. This can either be a path to a file name or uri containing the parameters declaration, or the parameters declaration as a string.";
         public const string SetPolicyDefinitionManagementGroupHelp = "The name of the management group of the policy definition to update.";
         public const string SetPolicyDefinitionSubscriptionIdHelp = "The subscription ID of the policy definition to update.";
+        public const string SetPolicyDefinitionInputObjectHelp = "The policy definition object to update that was output from another cmdlet.";
 
         /// <summary>
         /// Policy set definition cmdlet parameter help strings
@@ -100,6 +103,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation.Policy
         public const string NewPolicySetDefinitionSubscriptionIdHelp = "The subscription ID of the new policy set definition.";
         public const string RemovePolicySetDefinitionNameHelp = "The policy set definition name to delete.";
         public const string RemovePolicySetDefinitionIdHelp = "The fully qualified policy set definition ID to delete, including the subscription or management group. e.g. /subscriptions/{subscriptionId}/providers/Microsoft.Authorization/policySetDefinitions/{policySetDefinitionName}.";
+        public const string RemovePolicySetDefinitionInputObjectHelp = "The policy set definition object to remove that was output from another cmdlet.";
         public const string RemovePolicySetDefinitionManagementGroupHelp = "The name of the management group of the policy set definition to delete.";
         public const string RemovePolicySetDefinitionSubscriptionIdHelp = "The subscription ID of the policy set definition to delete.";
         public const string SetPolicySetDefinitionNameHelp = "The name of the policy set definition to update.";
@@ -111,5 +115,6 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation.Policy
         public const string SetPolicySetDefinitionParameterHelp = "The parameters declaration of the updated policy set definition. This can either be a path to a file name or uri containing the parameters declaration, or the parameters declaration as a string.";
         public const string SetPolicySetDefinitionManagementGroupHelp = "The name of the management group of the policy set definition to update.";
         public const string SetPolicySetDefinitionSubscriptionIdHelp = "The subscription ID of the policy set definition to update.";
+        public const string SetPolicySetDefinitionInputObjectHelp = "The policy set definition object to update that was output from another cmdlet.";
     }
 }